+++ lib/KTable/useSorting/index.js
